% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aneurisk65}
\alias{aneurisk65}
\title{Subset of the Aneurisk data set}
\format{
A list with abscissas x and values y:
\describe{
  \item{x}{Matrix 65x1380;}
  \item{y}{Array 65x3x1380.}
}
}
\usage{
aneurisk65
}
\description{
A data set containing 65 multi-dimensional curves from the Aneurisk project.
The curves represent the first derivatives of the 65 centerlines of the
internal carotid artery.
}
\keyword{datasets}
