% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-caps.R
\name{compare_caps}
\alias{compare_caps}
\title{Multiple CAPS \code{mcaps} objects}
\usage{
compare_caps(
  x,
  y,
  n_clusters_max = 5L,
  metric = c("l2", "pearson"),
  clustering_method = c("kmeans", "hclust-complete", "hclust-average", "hclust-single"),
  warping_class = c("affine", "dilation", "none", "shift", "srsf"),
  centroid_type = c("mean", "medoid")
)
}
\arguments{
\item{x}{A numeric matrix of shape \eqn{N \times M} specifying the grids of
size \eqn{M} on which each of the \eqn{N} curves have been observed.}

\item{y}{A numeric array of shape \eqn{N \times L \times M} specifying the
\eqn{N}-sample of \eqn{L}-dimensional curves observed on grids of size
\eqn{M}.}

\item{n_clusters_max}{An integer value specifying the maximum number of
clusers to use. Defaults to \code{5L}.}

\item{metric}{A string specifying the metric used to compare curves.
Choices are \code{"l2"} or \code{"pearson"}. Defaults to \code{"l2"}. This is used only
when \code{warping_class != "srsf"}.}

\item{clustering_method}{A character vector specifying one or more clustering
methods to be fit. Choices are \code{"kmeans"}, \code{"hclust-complete"},
\code{"hclust-average"} or \code{"hclust-single"}. Defaults to all of them.}

\item{warping_class}{A character vector specifying one or more classes of
warping functions to use for curve alignment. Choices are \code{"affine"},
\code{"dilation"}, \code{"none"}, \code{"shift"} or \code{"srsf"}. Defaults to all of them.}

\item{centroid_type}{A character vector specifying one or more ways to
compute centroids. Choices are \code{"mean"} or \code{"medoid"}. Defaults to all of
them.}
}
\value{
An object of class \code{mcaps} which is a \code{\link[tibble:tibble]{tibble::tibble}} storing the
objects of class \code{\link{caps}} in correspondence of each combination of possible
choices from the input arguments.
}
\description{
This function searches for clusters in the input data set using different
strategies and generates an object of class \code{mcaps} which stores multiple
objects of class \code{\link{caps}}. This is a helper function to facilitate comparison
of clustering methods and choice of an \emph{optimal} one.
}
\examples{
out <- compare_caps(
  x = simulated30$x,
  y = simulated30$y,
  n_clusters_max = 2,
  warping_class = "affine",
  clustering_method = "hclust-complete",
  centroid_type = "mean"
)
}
