% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdacluster-package.R
\docType{package}
\name{fdacluster}
\alias{fdacluster}
\title{K-mean alignment algorithm and variants for functional data}
\description{
The \strong{fdacluster} package allows to jointly perform clustering
and alignment of functional data.
}
\references{
\enumerate{
\item Sangalli, L.M., Secchi, P., Vantini, S. and Vitelli, V. (2010),
\href{https://www.sciencedirect.com/science/article/abs/pii/S0167947309004605}{K-mean alignment for curve clustering},
Computational Statistics and Data Analysis, 54, 1219-1233.
\item Sangalli, L.M., Secchi, P. and Vantini, S. (2014), \href{https://projecteuclid.org/journals/electronic-journal-of-statistics/volume-8/issue-2/Analysis-of-AneuRisk65-data-k-mean-alignment/10.1214/14-EJS938A.full}{Analysis of AneuRisk65 data: K-mean Alignment},
Electronic Journal of Statistics, 8 (2), 1891-1904.
}
}
