% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdadist.R
\name{fdadist}
\alias{fdadist}
\title{Distance Matrix for Functional Data with Amplitude and Phase Separation}
\usage{
fdadist(
  x,
  y,
  warping_class = c("affine", "dilation", "none", "shift", "srsf"),
  metric = c("l2", "pearson"),
  labels = NULL
)
}
\arguments{
\item{x}{A numeric matrix of shape \eqn{N \times M} specifying the grids of
size \eqn{M} on which each of the \eqn{N} curves have been observed.}

\item{y}{A numeric array of shape \eqn{N \times L \times M} specifying the
\eqn{N}-sample of \eqn{L}-dimensional curves observed on grids of size
\eqn{M}.}

\item{warping_class}{A string specifying the warping class Choices are
\code{"affine"}, \code{"dilation"}, \code{"none"}, \code{"shift"} or \code{"srsf"}. Defaults to
\code{"affine"}. The SRSF class is the only class which is boundary-preserving.}

\item{metric}{A string specifying the metric used to compare curves.
Choices are \code{"l2"} or \code{"pearson"}. Defaults to \code{"l2"}. This is used only
when \code{warping_class != "srsf"}.}

\item{labels}{A character vector specifying curve labels. Defaults to \code{NULL}
which uses sequential numbers as labels.}
}
\value{
A \link[stats:dist]{stats::dist} object storing the distance matrix between the input
curves using the metric specified through the argument \code{metric} and the
warping class specified by the argument \code{warping_class}.
}
\description{
This function computes the matrix of pairwise distances between curves a
functional data sample. This can be achieved with or without phase and
amplitude separation, which can be done using a variety of warping classes.
}
\examples{
D <- fdadist(simulated30$x, simulated30$y)
}
