% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdakmapp.R
\docType{package}
\name{fdakmapp}
\alias{fdakmapp}
\alias{fdakmapp-package}
\title{Functional Data Analysis Plus: K-Mean/Medoid Alignment:}
\description{
Fdakmap jointly performs clustering and alignment of a functional
dataset (multidimensional or unidimensional functions).
}
\examples{

}
\references{
\itemize{
\item Sangalli, L.M., Secchi, P., Vantini, S., Vitelli, V., 2010. \emph{"K-mean alignment for curve clustering"}. Computational Statistics and Data Analysis, 54, 1219-1233.
\item Sangalli, L.M., Secchi, P., Vantini, S., 2014. \emph{"Analysis of AneuRisk65 data: K-mean Alignment"}. Electronic Journal of Statistics, Special Section on "Statistics of Time Warpings and Phase Variations", Vol. 8, No. 2, 1891-1904.
}
}
\seealso{
\code{\link{kmap}}.
}
\keyword{Alignment,}
\keyword{Analysis,}
\keyword{Clustering,}
\keyword{Data}
\keyword{Functional}
\keyword{K-Mean}
\keyword{Registration,}
\keyword{Time}
\keyword{Warping}
