% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmaps-class.R
\name{kmaps}
\alias{kmaps}
\alias{as_kmaps}
\alias{is_kmaps}
\title{Class for K-Means with Amplitude and Phase Separation}
\usage{
as_kmaps(x)

is_kmaps(x)
}
\arguments{
\item{x}{A list coercible into an object of class \code{\link{kmaps}}.}
}
\description{
The k-means algorithm with joint amplitude and phase separation produces a
number of outputs. This class is meant to encapsulate them into a single
object for providing dedicated \code{S3} methods for e.g. plotting, summarizing,
etc. The name of the class stems from \strong{K}-\strong{M}eans with \strong{A}mplitude and
\strong{P}hase \strong{S}eparation.
}
\details{
An object of class \code{\link{kmaps}} is a list with the following components:
\itemize{
\item \code{original_curves}: A numeric matrix of shape \eqn{N \times L \times M}
storing a sample with the \eqn{N} \eqn{L}-dimensional original curves
observed on grids of size \eqn{M}.
\item \code{aligned_grids}: A numeric matrix of shape \eqn{N \times L \times M}
storing a sample with the \eqn{N} \eqn{L}-dimensional aligned curves observed
on grids of size \eqn{M}.
\item \code{center_curves}: A numeric matrix of shape \eqn{K \times L \times M}
storing the \eqn{K} centers which are \eqn{L}-dimensional curves observed on
a grid of size \eqn{M};
\item \code{grid}: A numeric vector of length \eqn{M} storing the
common grid of size \eqn{M} on which curves have been observed;
\item \code{n_clusters}: An integer value storing the number of clusters;
\item \code{memberships}: An integer vector of length \eqn{N} storing the cluster ID
which each curve belongs to;
\item \code{distances_to_center}: A numeric vector of length \eqn{N} storing the
distance of each curve to the center of its cluster;
\item \code{warpings}: A numeric matrix of shape \eqn{N \times M} storing the
estimated warping functions for each of the \eqn{N} curves evaluted on the
common \code{grid} of size \eqn{M};
\item \code{n_iterations}: An integer value storing the number of iterations
performed until convergence;
\item \code{call_name}: A string storing the name of the function that was used to
produce the k-means alignment results;
\item \code{call_args}: A list containing the exact arguments that were passed to
the function \code{call_name} that produced this output.
}
}
\examples{
res <- fdakmeans(
  simulated30$x,
  simulated30$y,
  seeds = c(1, 21),
  n_clusters = 2,
  centroid_type = "medoid",
  warping_class = "affine",
  distance = "pearson"
)
as_kmaps(res)
is_kmaps(res)
}
