% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kma_viz.R
\name{plot.kma}
\alias{plot.kma}
\title{Plot for \code{kmap} objects}
\usage{
\method{plot}{kma}(x, bp_sim = FALSE, wr_fun = FALSE, ...)
}
\arguments{
\item{x}{The \code{kmap} object to be plotted.}

\item{bp_sim}{Boolean. When \code{TRUE}, dissimilarity boxplot are
additionally plotted. Default is \code{FALSE}.}

\item{wr_fun}{Boolean. When \code{TRUE}, the warping functions are
additionally plotted. Default is \code{FALSE}.}

\item{...}{Additional parameters passed to the \code{\link[graphics]{plot}}
function.}
}
\value{
The input \code{kmap} object invisibly.
}
\description{
Plot for \code{kmap} objects
}
