% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmap_show_results.R
\name{plot.kmap}
\alias{plot.kmap}
\title{Plot for \code{kmap} objects}
\usage{
\method{plot}{kmap}(x, bp_sim = FALSE, wr_fun = FALSE, ...)
}
\arguments{
\item{x}{The \code{kmap} object to be plotted.}

\item{bp_sim}{Boolean. When \code{TRUE}, dissimilarity boxplot are
additionally plotted. Default is \code{FALSE}.}

\item{wr_fun}{Boolean. When \code{TRUE}, the warping functions are
additionally plotted. Default is \code{FALSE}.}
}
\value{
The input \code{kmap} object invisibly.
}
\description{
Plot for \code{kmap} objects
}
