% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim30_caps}
\alias{sim30_caps}
\title{A \code{caps} object from simulated data for examples}
\format{
An object of class \code{\link{caps}}.
}
\usage{
sim30_caps
}
\description{
An object of class \code{\link{caps}} storing the result of the \code{\link[=fdakmeans]{fdakmeans()}}
function applied on the data set \code{\link{simulated30}} using the affine warping
class and the Pearson metric and searching for 2 clusters.
}
\keyword{datasets}
