% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim30_mcaps}
\alias{sim30_mcaps}
\title{An \code{mcaps} object from simulated data for examples}
\format{
An object of class \code{mcaps} which is effectively a \link[tibble:tibble]{tibble::tibble}
with 5 columns and as many rows as there are clustering strategies to
compare. The 5 column-variables are:
\itemize{
\item \code{n_clusters}: The number of clusters;
\item \code{clustering_method}: The clustering method;
\item \code{warping_class}: The class of warping functions used for curve alignment;
\item \code{centroid_type}: The type of centroid used to compute a cluster
representative;
\item \code{caps_obj}: The result of the corresponding clustering strategy as
objects of class \code{\link{caps}}.
}
}
\usage{
sim30_mcaps
}
\description{
An object of class \code{mcaps} storing the result of the \code{\link[=compare_caps]{compare_caps()}}
function applied on the data set \code{\link{simulated30_sub}} for comparing the
clustering structures found by the \code{\link[=fdakmeans]{fdakmeans()}} function with \code{mean}
centroid type used with various classes of warping functions and varying
number of clusters.
}
\keyword{datasets}
