% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulated30}
\alias{simulated30}
\title{Data for Examples.}
\format{A list with abscissas x and values y:
\describe{
  \item{x}{matrix 30 x 200 }
  \item{y}{array 30 x 200 x 1}
}}
\usage{
simulated30
}
\description{
A dataset containing 30 simulated curves.
}
\keyword{datasets}
