% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulated30_sub}
\alias{simulated30_sub}
\title{Simulated data for examples}
\format{
A list with abscissas x and values y:
\describe{
\item{x}{Matrix 30x30;}
\item{y}{Array 30x1x30.}
}
}
\usage{
simulated30_sub
}
\description{
A data set containing 30 simulated uni-dimensional curves.
}
\keyword{datasets}
