% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulated90}
\alias{simulated90}
\title{Simulated data from the CSDA paper}
\format{
A list with abscissas x and values y:
\describe{
\item{x}{Vector of size 100;}
\item{y}{Matrix if size 90x100.}
}
}
\usage{
simulated90
}
\description{
A data set containing 90 simulated uni-dimensional curves.
}
\keyword{datasets}
