% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aneurisk65}
\alias{aneurisk65}
\title{Data for Examples.}
\format{A list with abscissas x and values y:
\describe{
  \item{x}{matrix 65 x 1380 }
  \item{y}{array 65 x1380 x 3}
}}
\usage{
aneurisk65
}
\description{
A dataset containing 65 multidimensional curves from aneurisk project.
The curves represent the first derivatives of the 65 ICA centerlines.
}
\keyword{datasets}
