% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caps-diagnostic.R
\name{diagnostic_plot}
\alias{diagnostic_plot}
\title{Diagnostic plot for the result of a clustering strategy stored in a \code{caps}
object}
\usage{
diagnostic_plot(x)
}
\arguments{
\item{x}{An object of class \code{\link{caps}}.}
}
\value{
An object of class \link[ggplot2:ggplot]{ggplot2::ggplot}.
}
\description{
This function plots the values of the distance to center and silhouette for
each observation. Observations are ordered within cluster by decreasing value
of silhouette.
}
\examples{
diagnostic_plot(sim30_caps)
}
