% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdakmapp.R
\docType{package}
\name{fdakmapp}
\alias{fdakmapp}
\alias{fdakmapp-package}
\title{K-mean alignment algorithm and variants for functional data}
\description{
The package \code{fdakmapp} allows to jointly perform clustering
and alignment of functional data.
}
\references{
\itemize{
\item Sangalli, L.M., Secchi, P., Vantini, S., Vitelli, V., 2010. \emph{"K-mean alignment for curve clustering"}. Computational Statistics and Data Analysis, 54, 1219-1233.
\item Sangalli, L.M., Secchi, P., Vantini, S., 2014. \emph{"Analysis of AneuRisk65 data: K-mean Alignment"}. Electronic Journal of Statistics, Special Section on "Statistics of Time Warpings and Phase Variations", Vol. 8, No. 2, 1891-1904.
}
}
