% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmap_show_results.R
\name{kmap_show_results}
\alias{kmap_show_results}
\title{plot results kmap}
\usage{
kma_show_results <-function (Result, bp_sim,wr_fun)
}
\arguments{
\item{Result}{output of kmap.}

\item{bp_sim}{boolean: if TRUE dissimilarity,similarity or distance boxplot are plotted. Default value is FALSE.}

\item{wr_fun}{boolean: if TRUE the warping functions applied to x are plotted. Default value is FALSE.}
}
\description{
Show results of the clustering with alignment of functional data
}
