% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kma_viz.R
\name{plot.kma}
\alias{plot.kma}
\title{Plot for \code{kmap} objects}
\usage{
\method{plot}{kma}(x, number_of_displayed_points = 50, ...)
}
\arguments{
\item{x}{The \code{kma} object to be plotted.}

\item{number_of_displayed_points}{The number of points to graphically
represents. It is set as the minimum between this parameter and the number
of points in the original data set. Defaults to 50.}

\item{...}{Additional parameters passed to the \code{\link[graphics]{plot}}
function.}
}
\value{
A \code{ggplot} object invisibly.
}
\description{
Plot for \code{kmap} objects
}
\examples{
res <- kma(
  simulated30$x,
  simulated30$y,
  seeds = c(1, 21),
  n_clust = 2,
  center_method = "medoid",
  warping_method = "affine",
  dissimilarity_method = "pearson"
)

plot(res)
}
