% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kma_viz.R
\name{plot.kma}
\alias{plot.kma}
\title{Plot for \code{kmap} objects}
\usage{
\method{plot}{kma}(x, bp_sim = FALSE, wr_fun = FALSE, ...)
}
\arguments{
\item{x}{The \code{kmap} object to be plotted.}

\item{bp_sim}{Boolean. When \code{TRUE}, dissimilarity boxplot are
additionally plotted. Default is \code{FALSE}.}

\item{wr_fun}{Boolean. When \code{TRUE}, the warping functions are
additionally plotted. Default is \code{FALSE}.}

\item{...}{Additional parameters passed to the \code{\link[graphics]{plot}}
function.}
}
\value{
The input \code{kmap} object invisibly.
}
\description{
Plot for \code{kmap} objects
}
\examples{
res <- kma(
  simulated30$x,
  simulated30$y,
  seeds = c(21, 13),
  n_clust = 2,
  center_method = "medoid",
  warping_method = "affine",
  dissimilarity_method = "pearson"
)

tmp1 <- res$y
tmp2 <- res$y_centers_final
res$y <- array(dim = c(dim(tmp1)[1], dim(tmp1)[3], dim(tmp1)[2]))
res$y_centers_final <- array(dim = c(dim(tmp2)[1], dim(tmp2)[3], dim(tmp2)[2]))
for (i in 1) {
  res$y[, , i] <- tmp1[, i, ]
  res$y_centers_final[, , i] <- tmp2[, i, ]
}

plot(res)
}
