% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulated30}
\alias{simulated30}
\title{Simulated data for examples}
\format{
A list with abscissas x and values y:
\describe{
\item{x}{Matrix 30x200;}
\item{y}{Array 30x1x200.}
}
}
\usage{
simulated30
}
\description{
A data set containing 30 simulated uni-dimensional curves.
}
\keyword{datasets}
